Fixed capitolization of directories
