--- main/src/Hex/Hex.java
@@ -1,48 +1,0 @@
-package Hex;
-
-import Terrain.Terrain;
-import Terrain.Rocky;
-
-
-/**
- * Created by christinemoore on 3/13/17.
- */
-public class Hex {
-    private int tileID;
-    private int settlementID;
-    private Terrain terrain;
-    private int level;
-
-    public Hex(){
-        // literally just made up a placeholder IDs until
-        // we decide on id format
-        tileID = (int)(Math.random() * 100);
-        settlementID = (int)(Math.random() * 100);
-        terrain = Rocky.getInstance();
-        level = 1;
-    }
-
-    public Hex(int tileID, int settlementID, Terrain terrain, int level){
-        this.tileID = tileID;
-        this.settlementID = settlementID;
-        this.terrain = terrain;
-        this.level = level;
-    }
-
-    public int getTileID(){
-        return tileID;
-    }
-
-    public int getSettlementID(){
-        return settlementID;
-    }
-
-    public Terrain getTerrain(){
-        return terrain;
-    }
-
-    public int getLevel(){
-        return level;
-    }
-
-}
