--- main/src/Piece/Piece.java
@@ -1,5 +1,0 @@
-package Piece;
-
-public interface Piece {
-    public void accept(PieceVisitor visitor);
-}
