--- main/src/Piece/Villager.java
@@ -1,10 +1,0 @@
-package Piece;
-
-/**
- * Created by mbregg on 3/15/17.
- */
-public class Villager implements Piece {
-    public void accept(PieceVisitor visitor) {
-        visitor.visitVillager(this);
-    }
-}
