--- main/src/Terrain/Grassland.java
@@ -1,16 +1,0 @@
-package Terrain;
-
-public class Grassland extends Terrain{
-	private static Grassland instance;
-	
-	private Grassland(){}
-	
-	public static Grassland getInstance(){
-		if(instance == null)
-		{
-			instance=new Grassland();
-		}
-		
-		return instance;
-	}
-}
