--- main/src/Terrain/Jungle.java
@@ -1,16 +1,0 @@
-package Terrain;
-
-public class Jungle extends Terrain{
-	private static Jungle instance;
-
-	private Jungle(){}
-
-	public static Jungle getInstance(){
-		if(instance == null)
-		{
-			instance=new Jungle();
-		}
-
-		return instance;
-	}
-}
