--- main/src/Terrain/Lake.java
@@ -1,16 +1,0 @@
-package Terrain;
-
-public class Lake extends Terrain {
-	private static Lake instance;
-
-	private Lake(){}
-
-	public static Lake getInstance(){
-		if(instance == null)
-		{
-			instance=new Lake();
-		}
-
-		return instance;
-	}
-}
