--- main/src/Terrain/Terrain.java
@@ -1,5 +1,0 @@
-package Terrain;
-
-public abstract class Terrain {
-
-}
