--- main/src/Terrain/Volcano.java
@@ -1,16 +1,0 @@
-package Terrain;
-
-public class Volcano extends Terrain {
-	private static Volcano instance;
-	
-	private Volcano(){}
-	
-	public static Volcano getInstance(){
-		if(instance == null)
-		{
-			instance=new Volcano();
-		}
-		
-		return instance;
-	}
-}
