+++ main/src/piece/Piece.java
@@ -1,0 +1,5 @@
+package piece;
+
+public interface Piece {
+    public void accept(PieceVisitor visitor);
+}
