+++ main/src/piece/PieceVisitor.java
@@ -1,0 +1,10 @@
+package piece;
+
+/**
+ * Created by mbregg on 3/15/17.
+ */
+public interface PieceVisitor {
+    public void visitVillager(Villager villager);
+
+    public void visitTotoro(Totoro totoro);
+}
