+++ main/src/piece/Totoro.java
@@ -1,0 +1,10 @@
+package piece;
+
+/**
+ * Created by mbregg on 3/15/17.
+ */
+public class Totoro implements Piece {
+    public void accept(PieceVisitor visitor) {
+        visitor.visitTotoro(this);
+    }
+}
