+++ main/src/terrain/Jungle.java
@@ -1,0 +1,16 @@
+package terrain;
+
+public class Jungle extends Terrain{
+	private static Jungle instance;
+
+	private Jungle(){}
+
+	public static Jungle getInstance(){
+		if(instance == null)
+		{
+			instance=new Jungle();
+		}
+
+		return instance;
+	}
+}
