+++ main/src/terrain/Lake.java
@@ -1,0 +1,16 @@
+package terrain;
+
+public class Lake extends Terrain {
+	private static Lake instance;
+
+	private Lake(){}
+
+	public static Lake getInstance(){
+		if(instance == null)
+		{
+			instance=new Lake();
+		}
+
+		return instance;
+	}
+}
