+++ main/src/terrain/Rocky.java
@@ -1,0 +1,16 @@
+package terrain;
+
+public class Rocky extends Terrain{
+	private static Rocky instance;
+	
+	private Rocky(){}
+	
+	public static Rocky getInstance(){
+		if(instance == null)
+		{
+			instance=new Rocky();
+		}
+		
+		return instance;
+	}
+}
