+++ main/src/terrain/Terrain.java
@@ -1,0 +1,5 @@
+package terrain;
+
+public abstract class Terrain {
+
+}
