+++ main/src/terrain/Volcano.java
@@ -1,0 +1,16 @@
+package terrain;
+
+public class Volcano extends Terrain {
+	private static Volcano instance;
+	
+	private Volcano(){}
+	
+	public static Volcano getInstance(){
+		if(instance == null)
+		{
+			instance=new Volcano();
+		}
+		
+		return instance;
+	}
+}
