+++ main/test/piece/TotoroTest.java
@@ -1,0 +1,21 @@
+package piece;
+
+import org.junit.Test;
+
+/**
+ * Created by mbregg on 3/15/17.
+ */
+public class TotoroTest {
+    @Test
+    public void acceptTest() throws Exception {
+        Totoro totoro = new Totoro();
+        totoro.accept(new PieceVisitor() {
+            public void visitTotoro(Totoro t) {
+                assert(t != null);
+            }
+            public void visitVillager(Villager v) {
+                assert(false);
+            }
+        });
+    }
+}
