+++ main/test/piece/VillagerTest.java
@@ -1,0 +1,21 @@
+package piece;
+
+import org.junit.Test;
+
+/**
+ * Created by mbregg on 3/15/17.
+ */
+public class VillagerTest {
+    @Test
+    public void acceptTest() throws Exception {
+        Villager villager = new Villager();
+        villager.accept(new PieceVisitor() {
+            public void visitVillager(Villager v) {
+                assert(v != null);
+            }
+            public void visitTotoro(Totoro t) {
+                assert(false);
+            }
+        });
+    }
+}
